Moved methods
They make more sense this way
