+++ BetterTriggers/Containers/References.cs
@@ -83,7 +83,7 @@ namespace BetterTriggers.Containers
 
             RemoveReferrer(t);
 
-            var parameters = Triggers.GetParametersFromTrigger(t);
+            var parameters = Parameter.GetParametersFromTrigger(t);
             var variables = Project.CurrentProject.Variables;
             var triggers = Project.CurrentProject.Triggers;
             parameters.ForEach(p =>
@@ -114,7 +114,7 @@ namespace BetterTriggers.Containers
             var triggers = Project.CurrentProject.Triggers.GetAll();
             triggers.ForEach(exTrig =>
             {
-                var parameters = Triggers.GetParametersFromTrigger(exTrig);
+                var parameters = Parameter.GetParametersFromTrigger(exTrig);
                 parameters.ForEach(p =>
                 {
                     if (p is VariableRef varRef)
