+++ BetterTriggers/Containers/Triggers.cs
@@ -193,32 +193,6 @@ namespace BetterTriggers.Containers
         }
 
 
-
-        /// <returns>A list of all parameters given to a TriggerElement</returns>
-        public static List<Parameter> GetElementParametersAll(TriggerElement te)
-        {
-            ECA eca = (ECA)te;
-            var list = GetElementParametersAll(eca.function.parameters);
-            return list;
-        }
-
-        private static List<Parameter> GetElementParametersAll(List<Parameter> parameters)
-        {
-            List<Parameter> list = new List<Parameter>();
-
-            for (int i = 0; i < parameters.Count; i++)
-            {
-                list.Add(parameters[i]);
-                if (parameters[i] is Function)
-                {
-                    Function f = (Function)parameters[i];
-                    list.AddRange(GetElementParametersAll(f.parameters));
-                }
-            }
-
-            return list;
-        }
-
         public void CopyTriggerElements(ExplorerElement copiedFrom, TriggerElementCollection copiedCollection, bool isCut = false)
         {
             var type = copiedCollection.Elements[0].ElementType;
@@ -239,288 +213,7 @@ namespace BetterTriggers.Containers
             else
                 CopiedElements.CutTriggerElements = null;
         }
-
-        /// <returns>A list of pasted elements.</returns>
-        public TriggerElementCollection PasteTriggerElements(ExplorerElement destinationTrigger, TriggerElement parentList, int insertIndex)
-        {
-            var copied = CopiedElements.CopiedTriggerElements;
-            var pasted = new TriggerElementCollection(copied.ElementType);
-            for (int i = 0; i < copied.Count(); i++)
-            {
-                if (copied.Elements[i] is ECA eca)
-                {
-                    pasted.Elements.Add(eca.Clone());
-                }
-                else if (copied.Elements[i] is LocalVariable localVar)
-                {
-                    var clone = localVar.Clone();
-                    var variables = Project.CurrentProject.Variables;
-                    clone.variable.Id = variables.GenerateId();
-                    clone.variable.Name = variables.GenerateLocalName(destinationTrigger.trigger, clone.variable.Name);
-                    clone.DisplayText = clone.variable.Name;
-                    pasted.Elements.Add(clone);
-                    variables.AddLocalVariable(clone);
-                }
-            }
-
-            if (CopiedElements.CutTriggerElements == null)
-            {
-                CommandTriggerElementPaste command = new CommandTriggerElementPaste(destinationTrigger, pasted, parentList, insertIndex);
-                command.Execute();
-            }
-            else
-            {
-                CommandTriggerElementCutPaste command = new CommandTriggerElementCutPaste(CopiedElements.CopiedFromTrigger, destinationTrigger, pasted, parentList, insertIndex);
-                command.Execute();
-            }
-
-            if(pasted.Elements.Count > 0)
-                pasted.Elements[pasted.Elements.Count - 1].IsSelected = true;
-
-            return pasted;
-        }
-
-
-        /// <returns>A list of every function in every trigger. This also includes inner functions in parameters.</returns>
-        public List<Function> GetFunctionsAll()
-        {
-            var triggers = GetAll();
-            List<Function> functions = new List<Function>();
-            triggers.ForEach(trig => functions.AddRange(GetFunctionsFromTrigger(trig)));
-
-            return functions;
-        }
-
-        public static List<Function> GetFunctionsFromTrigger(ExplorerElement explorerElement)
-        {
-            List<Function> list = new List<Function>();
-            list.AddRange(GatherFunctions(explorerElement.trigger.Events));
-            list.AddRange(GatherFunctions(explorerElement.trigger.Conditions));
-            list.AddRange(GatherFunctions(explorerElement.trigger.Actions));
-
-            return list;
-        }
-
-        private static List<Function> GatherFunctions(TriggerElementCollection triggerElements)
-        {
-            List<Function> list = new List<Function>();
-            triggerElements.Elements.ForEach(t =>
-            {
-                var eca = (ECA)t;
-                list.AddRange(GetFunctionsFromParameters(eca.function));
-
-                if (t is IfThenElse)
-                {
-                    var special = (IfThenElse)t;
-                    list.AddRange(GatherFunctions(special.If));
-                    list.AddRange(GatherFunctions(special.Then));
-                    list.AddRange(GatherFunctions(special.Else));
-                }
-                else if (t is AndMultiple)
-                {
-                    var special = (AndMultiple)t;
-                    list.AddRange(GatherFunctions(special.And));
-                }
-                else if (t is ForForceMultiple)
-                {
-                    var special = (ForForceMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is ForGroupMultiple)
-                {
-                    var special = (ForGroupMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is ForLoopAMultiple)
-                {
-                    var special = (ForLoopAMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is ForLoopBMultiple)
-                {
-                    var special = (ForLoopBMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is ForLoopVarMultiple)
-                {
-                    var special = (ForLoopVarMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is OrMultiple)
-                {
-                    var special = (OrMultiple)t;
-                    list.AddRange(GatherFunctions(special.Or));
-                }
-                else if (t is EnumDestructablesInRectAllMultiple)
-                {
-                    var special = (EnumDestructablesInRectAllMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is EnumDestructiblesInCircleBJMultiple)
-                {
-                    var special = (EnumDestructiblesInCircleBJMultiple)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-                else if (t is EnumItemsInRectBJ)
-                {
-                    var special = (EnumItemsInRectBJ)t;
-                    list.AddRange(GatherFunctions(special.Actions));
-                }
-            });
-
-            return list;
-        }
-
-        private static List<Function> GetFunctionsFromParameters(Function function)
-        {
-            List<Function> list = new List<Function>();
-            list.Add(function);
-            function.parameters.ForEach(p =>
-            {
-                if (p is VariableRef)
-                {
-                    VariableRef variableRef = p as VariableRef;
-                    Variable variable = Project.CurrentProject.Variables.GetById(variableRef.VariableId);
-                    if (variable == null)
-                        return;
-
-                    if (variable.IsArray)
-                    {
-                        if (variableRef.arrayIndexValues[0] is Function)
-                            list.AddRange(GetFunctionsFromParameters(variableRef.arrayIndexValues[0] as Function));
-                    }
-                    if (variable.IsTwoDimensions)
-                    {
-                        if (variableRef.arrayIndexValues[1] is Function)
-                            list.AddRange(GetFunctionsFromParameters(variableRef.arrayIndexValues[1] as Function));
-                    }
-                }
-
-                if (p is Function)
-                    list.AddRange(GetFunctionsFromParameters(p as Function));
-            });
-
-            return list;
-        }
-
-
-        /// <summary>
-        /// </summary>
-        /// <returns>A list of every parameter in every trigger.</returns>
-        public List<Parameter> GetParametersAll()
-        {
-            var triggers = GetAll();
-            List<Parameter> parameters = new List<Parameter>();
-            triggers.ForEach(trig => parameters.AddRange(GetParametersFromTrigger(trig)));
-
-            return parameters;
-        }
-
-        public List<VariableRef> GetVariableRefsFromTrigger(ExplorerElement explorerElement)
-        {
-            List<Parameter> _params = GetParametersFromTrigger(explorerElement);
-            List<VariableRef> variableRefs = new List<VariableRef>();
-            _params.ForEach(p =>
-            {
-                var varRef = p as VariableRef;
-                if (varRef != null)
-                    variableRefs.Add(varRef);
-            });
-
-            return variableRefs;
-        }
-
-        /// <summary>
-        /// </summary>
-        /// <returns>A list of every parameter in the given trigger.</returns>
-        public static List<Parameter> GetParametersFromTrigger(ExplorerElement explorerElement)
-        {
-            List<Parameter> list = new List<Parameter>();
-            list.AddRange(GatherTriggerParameters(explorerElement.trigger.Events));
-            list.AddRange(GatherTriggerParameters(explorerElement.trigger.Conditions));
-            list.AddRange(GatherTriggerParameters(explorerElement.trigger.LocalVariables));
-            list.AddRange(GatherTriggerParameters(explorerElement.trigger.Actions));
-
-            return list;
-        }
-
-        private static List<Parameter> GatherTriggerParameters(TriggerElementCollection triggerElements)
-        {
-            List<Parameter> parameters = new List<Parameter>();
-
-            for (int i = 0; i < triggerElements.Elements.Count; i++)
-            {
-                var triggerElement = triggerElements.Elements[i];
-                if (triggerElement is LocalVariable localVar)
-                {
-                    parameters.Add(localVar.variable.InitialValue);
-                    continue;
-                }
-
-                parameters.AddRange(GetElementParametersAll(triggerElement));
-
-                if (triggerElement is IfThenElse)
-                {
-                    var special = (IfThenElse)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.If));
-                    parameters.AddRange(GatherTriggerParameters(special.Then));
-                    parameters.AddRange(GatherTriggerParameters(special.Else));
-                }
-                else if (triggerElement is AndMultiple)
-                {
-                    var special = (AndMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.And));
-                }
-                else if (triggerElement is ForForceMultiple)
-                {
-                    var special = (ForForceMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is ForGroupMultiple)
-                {
-                    var special = (ForGroupMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is ForLoopAMultiple)
-                {
-                    var special = (ForLoopAMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is ForLoopBMultiple)
-                {
-                    var special = (ForLoopBMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is ForLoopVarMultiple)
-                {
-                    var special = (ForLoopVarMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is OrMultiple)
-                {
-                    var special = (OrMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Or));
-                }
-                else if (triggerElement is EnumDestructablesInRectAllMultiple)
-                {
-                    var special = (EnumDestructablesInRectAllMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is EnumDestructiblesInCircleBJMultiple)
-                {
-                    var special = (EnumDestructiblesInCircleBJMultiple)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-                else if (triggerElement is EnumItemsInRectBJ)
-                {
-                    var special = (EnumItemsInRectBJ)triggerElement;
-                    parameters.AddRange(GatherTriggerParameters(special.Actions));
-                }
-            }
-
-            return parameters;
-        }
-
+        
         public string GetValueName(string key, string returnType)
         {
             string text = key;
